Use a regular host with docker installed instead of a docker host
